♻️ ProblemScheduler로 문제 자동 생성 로직 이동
